tools/pyproject.toml
[project]
name = "tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv==1.2.2",
    "azure-identity==1.25.3",
    "azure-cognitiveservices-speech==1.49.0",
    "azure-ai-textanalytics==5.4.0",
    "requests==2.33.1",
    "azure-ai-documentintelligence==1.0.2",
]
